% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{addGroup}
\alias{addGroup}
\title{Add a Panel to a Study}
\usage{
addGroup(
  study_name,
  group_name,
  user_ids = c(),
  scrape_timelines = T,
  scrape_friends = T,
  scrape_followers = T,
  scrape_favorites = T,
  first_scrape = T,
  tokens,
  max_hours = 1,
  scrape_survey = F,
  qualtrics_survey_id = NULL
)
}
\arguments{
\item{study_name}{Name of the study to be added.}

\item{group_name}{Name of the panel to be added.}

\item{user_ids}{Twitter user ids of users to be added to the new panel.}

\item{scrape_timelines}{Should timelines be scraped for this panel?}

\item{scrape_friends}{Should friends be scraped for this panel?}

\item{scrape_followers}{Should followers be scraped for this panel?}

\item{scrape_favorites}{Should favorites be scraped for this panel?}

\item{first_scrape}{Should tricordR conduct an initial scrape of the specified data for the specified users?}

\item{tokens}{List of tokens to use for initial scrape.  See prepTokens.}

\item{max_hours}{Maximum duration of initial scrape.  Defaults to 1 hour.}

\item{scrape_survey}{Should tricordR conduct hourly scrapes of survey responses associated with this panel? Defaults to FALSE.}

\item{qualtrics_survey_id}{If scrape_survey is TRUE, provide a Qualtrics survey id to identify the relevant survey data.  Note: to use this, you must also install your Qualtrics authentication credentials on your machine.}
}
\description{
This function creates a new panel in a given study folder of tricordings, specifies desired data types to scrape, and optionally conducts an initial scrape of these data.
}
\examples{
addGroup()
}
\keyword{management}
