% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{addSentiment}
\alias{addSentiment}
\title{Add Sentiment}
\usage{
addSentiment(timelines_df)
}
\arguments{
\item{timelines_df}{Timeline dataframe to add sentiment scores to.}
}
\description{
A function to add sentiment scores to timeline data.
}
\examples{
addSentiment()
}
\keyword{sentiment}
