% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{addStudy}
\alias{addStudy}
\title{Add a Study}
\usage{
addStudy(study_name)
}
\arguments{
\item{study_name}{Name of the study to be added.}
}
\description{
This function creates a new study in the studies folder of tricordings, to which panels can be added.
}
\examples{
addStudy()
}
