% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{batchedSentiment}
\alias{batchedSentiment}
\title{Batched Sentiment}
\usage{
batchedSentiment(input, batch_size = 1000)
}
\arguments{
\item{input}{Text to classify for sentiment.}

\item{batch_size}{Size of batches to classify.  Defaults to 1000.}
}
\description{
A function to classify tweets for sentiment in batches.  Usually used via higher-level functions.
}
\examples{
batchedSentiment()
}
\keyword{sentiment}
