% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function2}
\alias{cat_function2}
\title{Another Cat Function}
\usage{
cat_function2(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats even more.
}
\examples{
cat_function2()
}
\keyword{cats}
