% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{chist}
\alias{chist}
\title{Centered Histogram}
\usage{
chist(x, b = 1)
}
\arguments{
\item{x}{Data to plot.}

\item{b}{Bandwidth, defaults to 1.}
}
\description{
Make a centered histogram breaks.
}
\examples{
chist()
}
\keyword{utility}
