% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{cSeq}
\alias{cSeq}
\title{Centered Sequence}
\usage{
cSeq(l = -2, u = 2, b = 0.1)
}
\arguments{
\item{l}{Lower bound}

\item{u}{Upper bound}

\item{b}{Bandwidth}
}
\description{
Make a centered sequence for histogram breaks
}
\examples{
cSeq()
}
\keyword{utility}
