% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{editGroup}
\alias{editGroup}
\title{Edit a Panel's Membership or Scrape Settings}
\usage{
editGroup(
  group_name,
  study_name,
  add_users = NULL,
  remove_users = NULL,
  scrape_timelines = NULL,
  scrape_friends = NULL,
  scrape_followers = NULL,
  scrape_favorites = NULL,
  first_scrape = FALSE,
  tokens = NULL,
  max_hours = 1,
  scrape_survey = NULL,
  qualtrics_survey_id = NULL
)
}
\arguments{
\item{group_name}{Name of the panel to be edited.}

\item{study_name}{Name of the study to be edited.}

\item{add_users}{Twitter user ids of users to be added to the panel.}

\item{remove_users}{Twitter user ids of users to be removed from the panel.}

\item{scrape_timelines}{Should timelines be scraped for this panel?}

\item{scrape_friends}{Should friends be scraped for this panel?}

\item{scrape_followers}{Should followers be scraped for this panel?}

\item{scrape_favorites}{Should favorites be scraped for this panel?}

\item{first_scrape}{Should tricordR conduct an initial scrape of the specified data for the specified users?}

\item{tokens}{List of tokens to use for first scrape.  See prepTokens.}

\item{max_hours}{Maximum duration of first scrape.  Defaults to 1 hour.}

\item{scrape_survey}{Should tricordR conduct hourly scrapes of survey responses associated with this panel? Defaults to FALSE.}

\item{qualtrics_survey_id}{If scrape_survey is TRUE, provide a Qualtrics survey id to identify the relevant survey data.  Note: to use this, you must also install your Qualtrics authentication credentials on your machine.}
}
\description{
This function creates a new panel in a given study folder of tricordings, specifies desired data types to scrape, and optionally conducts an initial scrape of these data.
}
\examples{
editGroup()
}
\keyword{management}
