% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_function.R
\name{external_function1}
\alias{external_function1}
\title{An external function}
\usage{
external_function1(x)
}
\arguments{
\item{x}{Put something in here}
}
\description{
An external function
}
\examples{
external_function1()
}
\keyword{cats}
