% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{firstScrape}
\alias{firstScrape}
\title{First Scrape of New Panel Members}
\usage{
firstScrape(
  user_ids,
  panel_directory,
  tokens,
  max_hours = 1,
  sentiment = "none",
  darmoc = FALSE
)
}
\arguments{
\item{user_ids}{A dataframe containing a user_id column and a penultimate_id column.}

\item{panel_directory}{The path to the panel folder to which users are being added.}

\item{tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 1 hour.}

\item{sentiment}{Should scraped tweets be analyzed for sentiment? Defaults to "none".}

\item{darmoc}{Should scrapted tweets be analyzed for ideology and sureness?  Defaults to FALSE}
}
\description{
This workhorse function conducts an inaugural scrape when new users are added to a panel.  It reads the scrape_settings saved in the panel folder, and scrapes the data types specified for that panel.
}
\examples{
firstScrape()
}
\keyword{scraping}
