% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{fixSentiment}
\alias{fixSentiment}
\title{Fix Sentiment}
\usage{
fixSentiment(input, allowed_range = c(-1, 1))
}
\arguments{
\item{input}{Sentiment values to fix}
}
\description{
A function to fix sentiment values.
}
\examples{
fixSentiment()
}
\keyword{dashboard}
