% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{get_friends_rotate_maxToken_BIG}
\alias{get_friends_rotate_maxToken_BIG}
\title{Get Friends (>5k)}
\usage{
get_friends_rotate_maxToken_BIG(users, n = 20000, list_tokens, max_hours = 1)
}
\arguments{
\item{users}{Either a character vector of user_ids, or a dataframe containing a user_id column.}

\item{n}{The maximum number of friends to scrape for each user.  Defaults to 20,000}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}
}
\description{
Scrapes friend lists for users, rotating through tokens, and using pagination to collect all friends, even when a user has more than 5,000 friends.
}
\examples{
get_friends_rotate_maxToken_BIG()
}
