% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{getFollowersBig}
\alias{getFollowersBig}
\title{Get Followers (>5k)}
\usage{
getFollowersBig(
  users,
  n = 20000,
  list_tokens,
  per_token_limit = 15,
  max_hours = 15
)
}
\arguments{
\item{users}{Either a character vector of user_ids, or a dataframe containing a user_id column.}

\item{n}{The maximum number of followers to scrape for each user.  Defaults to 20,000.}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{per_token_limit}{Batch size to reduce rate limit errors.  Defaults to 15 users per token.}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 1 hour.}
}
\description{
Scrapes follower lists for users, rotating through tokens, and detecting truncation due to rate limits, to collect all followers, even when a user has more than 5,000 followers.
}
\examples{
getFollowersBig()
}
\keyword{scraping}
