% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{getSampledUsersFriends}
\alias{getSampledUsersFriends}
\title{Friend Scraping for Sampled Users}
\usage{
getSampledUsersFriends(study_name, n = 250000, tokens, max_hours)
}
\arguments{
\item{study_name}{The name of the study folder to sample into.}

\item{n}{Maximum number of friends to scrape.  Defaults to 250,000.}

\item{tokens}{Tokens prepared using prepTokens().}

\item{max_hours}{Maximum time to spend scraping friends.}
}
\description{
This function scrapes users' friends, and is designed to work through a large volume robustly.
}
\examples{
getSampledUsersFriends()
}
\keyword{sampling}
