% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{getTimelinesHistorical}
\alias{getTimelinesHistorical}
\title{Get Historical Timelines (up to 3200 Tweets into the Past)}
\usage{
getTimelinesHistorical(
  users,
  n = 3200,
  list_tokens,
  per_token_limit = 100,
  max_hours = 4
)
}
\arguments{
\item{users}{A dataframe containing a user_id column, or a character vector of user ids.}

\item{n}{The maximum number of followers to scrape for each user.  Defaults to 3,200.}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{per_token_limit}{Batch size to reduce rate limit errors.  Defaults to 100 users per token.}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 4 hours.}
}
\description{
The workhorse function of historical timeline collection, takes a dataframe of containing a user_id column, and scrapes as many tweets into the past as specified, up to 3200.  This function is only sometimes used on its own, but plays an important part in higher-level scraping functions.
}
\examples{
getTimelinesHistorical()
}
\keyword{scraping}
