% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{initialize}
\alias{initialize}
\title{Initialization Function}
\usage{
initialize(location = "~/")
}
\arguments{
\item{location}{Location of tricordings directory, defaults to home directory (~/) and should not be changed at this time.}
}
\description{
This function initializes tricordR by creating the tricordings directory to save data, tokens, and scraping logs.
}
\examples{
initialize()
}
