% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{linePlot}
\alias{linePlot}
\title{Line Plot}
\usage{
linePlot(data_e, days, volume_smoothing, axis_cex)
}
\arguments{
\item{data_e}{timeline data to plot}
}
\description{
A function to visualize timeline scrapes as line plots.
}
\examples{
linePlot()
}
\keyword{dashboard}
