% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{link_ids}
\alias{link_ids}
\title{Link IDs}
\usage{
link_ids(claims, new_friends_by_follower)
}
\arguments{
\item{claims}{Accounts claimed by the participant.}

\item{new_friends_by_follower}{New friends by follower.}
}
\description{
A function to link ids.
}
\examples{
link_ids()
}
\keyword{matching}
