% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{loadTimelines}
\alias{loadTimelines}
\title{Load Timelines}
\usage{
loadTimelines(study_name, panel_name)
}
\arguments{
\item{study_name}{The name of the study in which the panel to load is located.}

\item{panel_name}{The name of the panel to load.}
}
\description{
A user-friendly function to load all timeline data collected for a given panel.
}
\examples{
loadTimelines()
}
\keyword{dashboard}
