% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{logToken}
\alias{logToken}
\title{Log Token Usage}
\usage{
logToken(list_tokens, token_index, scrape_object, scraping_function, user_id)
}
\arguments{
\item{list_tokens}{List of tokens in use by higher-level function.}

\item{token_index}{Index of token in current use.}

\item{scrape_object}{Object that should have just been scraped}

\item{scraping_function}{Name of higher-level function using tokens.}

\item{user_id}{Id of user attempting to scrape.}
}
\description{
This function logs token usage throughout scraping operations.
}
\examples{
logToken()
}
\keyword{scraping}
