% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{make_survey_timeseries}
\alias{make_survey_timeseries}
\title{make_survey_timeseries}
\usage{
make_survey_timeseries(
  input,
  volume_smoothing,
  survey_data_joined_GOOD,
  survey_start_date
)
}
\description{
A function to make_survey_timeseries
}
\examples{
make_survey_timeseries()
}
\keyword{dashboard}
