% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{make_timeseries}
\alias{make_timeseries}
\title{Make Time Series}
\usage{
make_timeseries(input, volume_smoothing, midnight_today, time_range)
}
\arguments{
\item{input}{Input data to reformat as time series.}
}
\description{
A function to make time series data.
}
\examples{
make_timeseries()
}
\keyword{dashboard}
