% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{makeSentiment}
\alias{makeSentiment}
\title{Make Sentiment}
\usage{
makeSentiment(input)
}
\arguments{
\item{input}{Text to classify for sentiment.}
}
\description{
A function to classify tweets for sentiment.  Usually used via higher-level functions.
}
\examples{
makeSentiment()
}
\keyword{sentiment}
