% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{match_async_by_time}
\alias{match_async_by_time}
\title{Match Asynchronously by Survey Start and End Time}
\usage{
match_async_by_time(
  responses_new,
  study_name,
  panel_name,
  assignment_panel,
  participant_tokens = NULL,
  this_timecode = NULL,
  add = TRUE
)
}
\arguments{
\item{responses_new}{New survey responses}

\item{study_name}{Name of study}

\item{panel_name}{Name of participant panel, passed from higher-level function.}

\item{assignment_panel}{Name of the assignment accounts panel, passed from higher-level function.}

\item{participant_tokens}{Tokens to scrape participant twitter data, passed from higher-level function.}

\item{this_timecode}{Timecode passed from higher-level function}

\item{add}{Go ahead and add these users to the participants panel? Defaults to TRUE.}
}
\description{
A function to match survey responses to twitter IDs based on following assigned accounts, using the survey times to be more precise.  Flexibly handles up to 999 assignments, by detecting the number of columns named "followNUMERAL" and "fNUMERAL" where NUMERAL is an integer between 1 and 999.
}
\examples{
match_async_by_time()
}
\keyword{matching}
