% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{match_by_following_3}
\alias{match_by_following_3}
\title{Match by Following 3}
\usage{
match_by_following_3(
  responses_new,
  study_name,
  panel_name,
  assignment_dir,
  participant_tokens,
  this_timecode
)
}
\arguments{
\item{responses_new}{New survey responses}

\item{study_name}{Name of study}

\item{panel_name}{Name of participant panel, passed from higher-level function.}

\item{assignment_dir}{Directory of the assignment accounts, passed from higher-level function.}

\item{participant_tokens}{Tokens to scrape participant twitter data, passed from higher-level function.}

\item{this_timecode}{Timecode passed from higher-level function}
}
\description{
A function to match survey responses to twitter IDs based on following assigned accounts
}
\examples{
match_by_following_3()
}
\keyword{matching}
