% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{match_by_following_3_INVESTIGATE}
\alias{match_by_following_3_INVESTIGATE}
\title{Match by Following 3 (Investigation Function)}
\usage{
match_by_following_3_INVESTIGATE(
  responses_new,
  before,
  after,
  study_name,
  panel_name = "participants",
  assignment_panel = "assignments",
  add = FALSE,
  tokens = NULL,
  use_claims = FALSE
)
}
\arguments{
\item{responses_new}{New survey responses}

\item{before}{Follower dataframe selected to represent the "before survey response" period.}

\item{after}{Follower dataframe selected to represent the "after survey response" period.}

\item{study_name}{Name of study}

\item{panel_name}{Name of participant panel, passed from higher-level function.  Defaults to "participants".}

\item{assignment_panel}{Name of participant panel, passed from higher-level function.  Defaults to "assignments".}

\item{add}{Go ahead and add these users to the panel?  Defaults to FALSE.}

\item{tokens}{Tokens to scrape participant twitter data, passed from higher-level function.}

\item{use_claims}{Condition matching on respondent claims?  Defaults to FALSE.}
}
\description{
A function to match survey responses to twitter IDs based on following assigned accounts, for investigation.
}
\examples{
match_by_following_3_INVESTIGATE()
}
\keyword{matching}
