% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{match_byhand}
\alias{match_byhand}
\title{Match by Hand}
\usage{
match_byhand(
  responses_new,
  user_id,
  study_name,
  panel_name = "participants",
  assignment_panel = "assignments",
  add = FALSE,
  tokens = NULL
)
}
\arguments{
\item{responses_new}{New survey responses}

\item{user_id}{User ID, manually entered, to associated with the new response.}

\item{study_name}{Name of study}

\item{panel_name}{Name of participant panel, passed from higher-level function.  Defaults to "participants".}

\item{assignment_panel}{Name of participant panel, passed from higher-level function.  Defaults to "assignments".}

\item{add}{Go ahead and add these users to the panel?  Defaults to FALSE.}

\item{tokens}{Tokens to scrape participant twitter data, passed from higher-level function.}
}
\description{
A function to match survey responses to twitter IDs based on following assigned accounts, for manual remediation.  Can handle up to 999 assignment accounts.
}
\examples{
match_byhand()
}
\keyword{matching}
