% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{maxN}
\alias{maxN}
\title{Return Nth Max}
\usage{
maxN(x, N = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{N}{An integer specifying which element to return. Defaults to 2, returning the second-greatest element.}
}
\description{
Order an input by numeric value and return the Nth largest element
}
\examples{
maxN()
}
