% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{maxNchar}
\alias{maxNchar}
\title{Return Nth Max of Character Vector}
\usage{
maxNchar(x, N = 2)
}
\arguments{
\item{x}{Character vector}

\item{N}{An integer specifying which element to return. Defaults to 2, returning the second-greatest element.}
}
\description{
Order a character input by alphanumeric value and return the Nth largest element as a character string
}
\examples{
maxNchar()
}
\keyword{utility}
