% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{myIndexer}
\alias{myIndexer}
\title{Indexer}
\usage{
myIndexer(input, reference)
}
\arguments{
\item{input}{Input values to index}

\item{reference}{Reference values against which to index}
}
\description{
A function to make index values.
}
\examples{
myIndexer()
}
\keyword{dashboard}
