% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{myRescale}
\alias{myRescale}
\title{Rescale Variable}
\usage{
myRescale(input)
}
\arguments{
\item{input}{Numeric input to rescale}
}
\description{
A function to rescale a numeric input to the 0-1 interval.
}
\examples{
myRescale()
}
\keyword{utility}
