% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{nodeIndexer}
\alias{nodeIndexer}
\title{Index Network Nodes}
\usage{
nodeIndexer(twitter_user_id, metadata)
}
\arguments{
\item{twitter_user_id}{The twitter user_id to look up.}

\item{metadata}{The metadata in which to look up the twitter user_id.}
}
\description{
A function to index network nodes for visualization
}
\examples{
nodeIndexer()
}
\keyword{dashboard}
