% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{plotGradient}
\alias{plotGradient}
\title{Gradients for Plots}
\usage{
plotGradient(
  input,
  left_color = c(0, 0, 1, 0.5),
  right_color = c(1, 0, 0, 0.5)
)
}
\arguments{
\item{input}{Input values to map to a color spectrum}
}
\description{
A function to make color gradients for plots.
}
\examples{
plotGradient()
}
\keyword{graphics}
