% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{prep_survey_data}
\alias{prep_survey_data}
\title{Prep survey data}
\usage{
prep_survey_data(experiment_directory, panel_name)
}
\description{
A function to prepare survey data for a dashboard.
}
\examples{
prep_survey_data()
}
\keyword{dashboard}
