% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{prep_timeline_data}
\alias{prep_timeline_data}
\title{Prep Timeline Data}
\usage{
prep_timeline_data(
  panel_directory,
  sessions_back,
  include_historical = FALSE,
  load_all_since_first = FALSE
)
}
\arguments{
\item{panel_directory}{The directory in which panel data is stored.}
}
\description{
A function to read timeline data into R.
}
\examples{
prep_timeline_data()
}
\keyword{dashboard}
