% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{prepTokens}
\alias{prepTokens}
\title{Prepare Tokens for Cycling}
\usage{
prepTokens(tokenset, which_tokens = 1:9)
}
\arguments{
\item{tokenset}{Name of token set to load}

\item{which_tokens}{A numeric vector indicating which tokens to load from the set.  Defaults to 1:9, loading all tokens.}
}
\description{
This function loads tokens as a list, for use by cycling functions
}
\examples{
prepTokens()
}
