% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{read_and_session}
\alias{read_and_session}
\title{Read and Session}
\usage{
read_and_session(input)
}
\arguments{
\item{input}{Directory from which to read scrapes}
}
\description{
A function to read timeline scrapes and append their session ids.
}
\examples{
read_and_session()
}
\keyword{dashboard}
