% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{readRelevantAssignmentFollowers}
\alias{readRelevantAssignmentFollowers}
\title{Read in Relevant Followers Only}
\usage{
readRelevantAssignmentFollowers(path, relevant_user_ids)
}
\arguments{
\item{path}{Path to RDS file to read}

\item{relevant_user_ids}{Set of relevant user_ids to keep}
}
\description{
A workhorse function to prepare network data for higher-level functions
}
\examples{
readRelevantAssignmentFollowers()
}
\keyword{dashboard}
