% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{runDash}
\alias{runDash}
\title{Run Dashboard}
\usage{
runDash(dashboard, port = 4201)
}
\arguments{
\item{dashboard}{Which Dashboard to run.}

\item{port}{Specify the port to run the dashboard on.  Defaults to 4201.}
}
\description{
A convenient function to launch any dashboard.
}
\examples{
runDash()
}
\keyword{dashboard}
