% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{runNetworkDash}
\alias{runNetworkDash}
\title{Run Network Dashboard}
\usage{
runNetworkDash(port = 4202)
}
\arguments{
\item{port}{Specify the port to run the dashboard on.  Defaults to 4202.}
}
\description{
A convenient function to launch the interactive network graph dashboard.
}
\examples{
runNetworkDash()
}
\keyword{dashboard}
