% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{runTimelineDashD3}
\alias{runTimelineDashD3}
\title{Run Timeline Dashboard}
\usage{
runTimelineDashD3(port = 4202)
}
\arguments{
\item{port}{Specify the port to run the dashboard on.  Defaults to 4201.}
}
\description{
A convenient function to launch the interactive timeline data collection dashboard.
}
\examples{
runTimelineDashD3()
}
\keyword{dashboard}
