% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{sampleEnglishTweeters5k50kRatioHalf}
\alias{sampleEnglishTweeters5k50kRatioHalf}
\title{Tweeter Sampling (English Language, between 5k 50k followers, friend/foll ratio<.5)}
\usage{
sampleEnglishTweeters5k50kRatioHalf(study_name, token, minutes = 1)
}
\arguments{
\item{study_name}{The name of the study folder to sample into.}

\item{token}{A single rtweet token.}

\item{minutes}{Minutes to keep sampling.}
}
\description{
This function samples tweets for a specified time duration (in 10-second chunks), then filters the resulting tweets to identify users tweeting in english and with at least 5000 followers.  Then it saves the unique user ids in the standard directory.
}
\examples{
sampleEnglishTweeters5k50kRatioHalf()
}
\keyword{sampling}
