% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{saveToken}
\alias{saveToken}
\title{Save Tokens for Cycling Functions}
\usage{
saveToken(set_name, consumer_key, consumer_secret, access_token, access_secret)
}
\arguments{
\item{set_name}{Name of token set (sets should correspond to Twitter developer accounts)}

\item{consumer_key}{Your app's consumer key string}

\item{consumer_secret}{Your app's consumer secret string}

\item{access_token}{Your app's access token string}

\item{access_secret}{Your app's access secret string}
}
\description{
This function initializes tricordR by creating the tricordings directory to save data, tokens, and scraping logs.
}
\examples{
saveToken()
}
