% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{scrapeFollowers}
\alias{scrapeFollowers}
\title{Scrape Followers of a Panel}
\usage{
scrapeFollowers(
  panel_directory,
  list_tokens,
  n = 20000,
  per_token_limit = 15,
  max_hours = 8
)
}
\arguments{
\item{panel_directory}{The path to the panel folder corresponding to the set of users to be scraped.}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{n}{The maximum number of followers to scrape for each user.  Defaults to 20,000.}

\item{per_token_limit}{Batch size to reduce rate limit errors.  Defaults to 15 users per token.}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 1 hour.}
}
\description{
A high-level function to scrape followers of users in a given panel, and save the resulting data in that panel folder.
}
\examples{
scrapeFollowers()
}
\keyword{scraping}
