% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{scrapeQualtrics}
\alias{scrapeQualtrics}
\title{Scrape Qualtrics}
\usage{
scrapeQualtrics(
  study_name,
  panel_name,
  match_by = NULL,
  assignment_panel = NULL,
  max_treat_followers = 60000,
  treatment_tokens,
  participant_tokens,
  add = FALSE
)
}
\arguments{
\item{study_name}{Name of study.}

\item{panel_name}{Name of participant panel to scrape survey data from.}

\item{match_by}{Matching algorithm to use, if applicable.}

\item{max_treat_followers}{Maximum number of followers to attempt to scrape from treatments. Defaults to 60,000.}

\item{treatment_tokens}{Tokens to scrape treatment followers.}

\item{participant_tokens}{Tokens to scrape participant data.}

\item{add}{Should new matches be added to the participants panel?  Defaults to FALSE, for safety.}

\item{assignment_dir}{Directory of assignment panel.}
}
\description{
A high-level qualtrics scraping function.
}
\examples{
scrapeQualtrics()
}
\keyword{matching}
