% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{scrapeStudy}
\alias{scrapeStudy}
\title{Scrape All Specified Data for All Panels in a Study}
\usage{
scrapeStudy(
  study_name,
  tokens,
  include_timelines = TRUE,
  include_friends = TRUE,
  include_followers = TRUE,
  include_favorites = TRUE,
  sentiment = "sentimentR",
  lasso = TRUE
)
}
\arguments{
\item{study_name}{The path to the panel folder corresponding to the set of users to be scraped.}

\item{tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{include_timelines}{Should timelines be scraped?}

\item{include_friends}{Should timelines be scraped?}

\item{include_followers}{Should timelines be scraped?}

\item{include_favorites}{Should favorites be scraped?}

\item{sentiment}{Should tweets be analyzed for sentiment?  Defaults to sentimentR.}

\item{lasso}{Should tweets be analyzed for ideology and sureness? Defaults to TRUE.}
}
\description{
A high-level function to scrape all data types specified in scrape_settings for users in all panel in a given study, and save the resulting data in the panel folders.
}
\examples{
scrapeStudy()
}
\keyword{scraping}
