% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{scrapeTimelines}
\alias{scrapeTimelines}
\title{Scrape Timelines of a Panel}
\usage{
scrapeTimelines(
  panel_directory,
  N = 3200,
  list_tokens,
  max_hours = 12,
  allHistory = FALSE,
  sub_batch_size = 100,
  sentiment = FALSE,
  darmoc = FALSE
)
}
\arguments{
\item{panel_directory}{The path to the panel folder corresponding to the set of users to be scraped.}

\item{N}{The number of tweets to be scraped.  Defaults to 3200, the maximum.}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 12 hours.}

\item{sub_batch_size}{Batch size to reduce rate limit errors.  Defaults to 100 users per token.}

\item{sentiment}{Should tweets be analyzed for sentiment?  Defaults to FALSE.}

\item{darmoc}{Should tweets be analyzed for ideology and sureness? Defaults to FALSE.}
}
\description{
A high-level function to scrape timelines of users in a given panel, and save the resulting data in that panel folder.
}
\examples{
scrapeTimelines()
}
\keyword{scraping}
