% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{singleNodeIndexer}
\alias{singleNodeIndexer}
\title{Index Single Network Nodes}
\usage{
singleNodeIndexer(twitter_user_id, thismetadata)
}
\arguments{
\item{twitter_user_id}{The twitter user_id to look up.}

\item{thismetadata}{The metadata in which to look up the twitter user_id.}
}
\description{
A workhorse function for nodeIndexer
}
\examples{
singleNodeIndexer()
}
\keyword{dashboard}
