% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{sn_to_userid}
\alias{sn_to_userid}
\title{Match Screenname to User ID}
\usage{
sn_to_userid(sn, treatment_acct_info)
}
\arguments{
\item{sn}{Screen name to match}

\item{treatment_acct_info}{Reference info set to consult.}
}
\description{
A function to match screen names to user ids.
}
\examples{
sn_to_userid()
}
\keyword{utility}
