% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{test_function1}
\alias{test_function1}
\title{A test function}
\usage{
test_function1(x)
}
\arguments{
\item{x}{Put something in here}
}
\description{
A test function
}
\examples{
test_function1()
}
\keyword{cats}
