% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{test_function2}
\alias{test_function2}
\title{A function that uses the test function}
\usage{
test_function2(x)
}
\arguments{
\item{x}{Put something in here}
}
\description{
A function that uses the test function
}
\examples{
test_function2()
}
\keyword{cats}
