% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{test_function3}
\alias{test_function3}
\title{A function that uses the external function}
\usage{
test_function3(x)
}
\arguments{
\item{x}{Put something in here}
}
\description{
A function that uses the external function
}
\examples{
test_function3()
}
\keyword{cats}
