% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{ts_to_cumulative}
\alias{ts_to_cumulative}
\title{ts_to_cumulative}
\usage{
ts_to_cumulative(survey_ts)
}
\description{
A function to convert time series data to cumulative data
}
\examples{
ts_to_cumulative()
}
\keyword{dashboard}
