% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{updateTimelines}
\alias{updateTimelines}
\title{Update Timelines}
\usage{
updateTimelines(
  users_df,
  n = 3200,
  list_tokens,
  per_token_limit = 100,
  max_hours = 4
)
}
\arguments{
\item{users_df}{A dataframe containing a user_id column and a penultimate_id column.}

\item{n}{The maximum number of followers to scrape for each user.  Defaults to 3,200.}

\item{list_tokens}{The list of tokens to be used for scraping.  See prepTokens().}

\item{per_token_limit}{Batch size to reduce rate limit errors.  Defaults to 100 users per token.}

\item{max_hours}{The maximum number of hours to continue scraping.  Defaults to 4 hours.}
}
\description{
The workhorse function of prospective timeline collection, takes a dataframe of user ids and penultimate tweet ids, and scrapes all new tweets since the penultimate tweet.  Penultimate tweets are used to avoid any ambiguity about whether all tweets have been collected.  This function is not generally used on its own, but plays an important part in higher-level scraping functions.
}
\examples{
updateTimelines()
}
\keyword{scraping}
