% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{which_maxes}
\alias{which_maxes}
\title{Return index of maximum value}
\usage{
which_maxes(x)
}
\arguments{
\item{x}{Input vector to maximize.}
}
\description{
A function to return the index of the maximum value in the input.
}
\examples{
which_maxes()
}
\keyword{utility}
