% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray.R
\name{xray}
\alias{xray}
\title{Mutate using X-rays}
\usage{
xray(.data, ..., .log = NULL)
}
\arguments{
\item{.data}{tibble}

\item{...}{arguments to dplyr::mutate}

\item{.log}{logging object}
}
\value{
tibble
}
\description{
Mutate using X-rays
}
\examples{
xray(dplyr::starwars, mass = base::round(mass, -1))
}
