% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{dotPlot}
\alias{dotPlot}
\title{Dot Plot}
\usage{
dotPlot(
  data_e,
  data_e_f,
  days,
  color_variable,
  show_names,
  sentiment_left_color,
  sentiment_right_color,
  ideo_left_color,
  ideo_right_color,
  point_cex,
  default_axis_cex,
  default_screen_name_cex,
  sentiment_reference_scale = NA,
  ideo_reference_scale = NA
)
}
\arguments{
\item{data_e}{Timeline data to plot.}

\item{data_e_f}{Timeline metadata to plot.}
}
\description{
A function to visualize timeline scrapes as dot plots.
}
\examples{
dotPlot()
}
\keyword{dashboard}
