% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLFUNCTIONS.R
\name{prep_network_data_d3}
\alias{prep_network_data_d3}
\title{Prepare Network Data for D3 Visualization}
\usage{
prep_network_data_d3(experiment_directory, participant_panel, assignment_panel)
}
\arguments{
\item{experiment_directory}{Directory to visualize}

\item{participant_panel}{Name of the panel representing "participants".}

\item{assignment_panel}{Name of the panel representing "assignments".}
}
\description{
A function to prepare network data for D3 visualization in dashboards.
}
\examples{
prep_network_data_d3()
}
\keyword{dashboard}
